+++ src/Yaat.Client/Views/Radar/RadarView.axaml.cs
@@ -512,6 +512,7 @@ public partial class RadarView : UserControl
         _canvas.EuroScopeMode = prefs.EuroScopeMode;
         _canvas.FlashNoLandingClearance = prefs.FlashNoLandingClearance;
         _canvas.ShowSpeechBubbles = prefs.ShowSpeechBubbles;
+        _canvas.AlwaysShowGroundBubblesOnRadar = prefs.AlwaysShowGroundBubblesOnRadar;
         _canvas.DatablockTextSize = prefs.RadarDatablockFontSize;
         Flyouts.FlyoutAppearance.FontSize = prefs.RadarFlyoutFontSize;
     }